requirements.txt
torch==2.9.1
diffusers==0.36.0
transformers==4.57.3
accelerate==1.12.0
pydantic==2.12.4
openai==2.11.0
anthropic==0.75.0
langchain==1.1.3
langchain-openai==1.1.3
langchain-anthropic==1.3.0
gradio==6.1.0
pillow==12.0.0
sentencepiece==0.2.1
protobuf==6.33.2
python-dotenv==1.2.1

